scripts/record-hike/requirements.txt
pydantic>=2.0
requests
python-dotenv
pyyaml
tabulate
rich
python-dateutil
